+++ tests/connection_load_balancing.rs
@@ -7,6 +7,7 @@ use russh::{
 };
 use russh_keys::{key::PrivateKeyWithHashAlg, load_secret_key};
 use sandhole::{entrypoint, ApplicationConfig, BindHostnames, LoadBalancing};
+use ssh_key::HashAlg;
 use tokio::{
     io::AsyncReadExt,
     net::TcpStream,
@@ -14,7 +15,7 @@ use tokio::{
 };
 
 #[tokio::test(flavor = "multi_thread")]
-async fn tcp_allow_requested_ports() {
+async fn tcp_load_balancing() {
     // 1. Initialize Sandhole
     let config = ApplicationConfig {
         domain: "foobar.tld".into(),
@@ -60,41 +61,63 @@ async fn tcp_allow_requested_ports() {
         panic!("Timeout waiting for Sandhole to start.")
     };
 
-    // 2. Start SSH client that will be proxied
-    let key = load_secret_key(
+    // 2. Start SSH clients that will be load-balanced
+    let key_1 = load_secret_key(
         concat!(env!("CARGO_MANIFEST_DIR"), "/tests/data/private_keys/key1"),
         None,
     )
     .expect("Missing file key1");
-    let ssh_client = SshClient;
-    let mut session = russh::client::connect(Default::default(), "127.0.0.1:18022", ssh_client)
+    let ssh_client_a = SshClientA;
+    let mut session_a = russh::client::connect(Default::default(), "127.0.0.1:18022", ssh_client_a)
         .await
         .expect("Failed to connect to SSH server");
-    assert!(session
+    assert!(session_a
         .authenticate_publickey(
             "user",
-            PrivateKeyWithHashAlg::new(Arc::new(key), None).unwrap()
+            PrivateKeyWithHashAlg::new(Arc::new(key_1), None).unwrap()
         )
         .await
         .expect("SSH authentication failed"));
-    session
+    session_a
+        .tcpip_forward("localhost", 12345)
+        .await
+        .expect("tcpip_forward failed");
+    let key_2 = load_secret_key(
+        concat!(env!("CARGO_MANIFEST_DIR"), "/tests/data/private_keys/key2"),
+        None,
+    )
+    .expect("Missing file key1");
+    let ssh_client_b = SshClientB;
+    let mut session_b = russh::client::connect(Default::default(), "127.0.0.1:18022", ssh_client_b)
+        .await
+        .expect("Failed to connect to SSH server");
+    assert!(session_b
+        .authenticate_publickey(
+            "user",
+            PrivateKeyWithHashAlg::new(Arc::new(key_2), Some(HashAlg::Sha512)).unwrap()
+        )
+        .await
+        .expect("SSH authentication failed"));
+    session_b
         .tcpip_forward("localhost", 12345)
         .await
         .expect("tcpip_forward failed");
 
     // 3. Connect to the TCP port of our proxy
-    let mut tcp_stream = TcpStream::connect("127.0.0.1:12345")
-        .await
-        .expect("TCP connection failed");
-    let mut buf = String::with_capacity(13);
-    tcp_stream.read_to_string(&mut buf).await.unwrap();
-    assert_eq!(buf, "Hello, world!");
+    for _ in 0..10 {
+        let mut tcp_stream = TcpStream::connect("127.0.0.1:12345")
+            .await
+            .expect("TCP connection failed");
+        let mut buf = Vec::with_capacity(3);
+        tcp_stream.read_to_end(&mut buf).await.unwrap();
+        assert!(matches!(&buf[..], b"A" | b"B"));
+    }
 }
 
-struct SshClient;
+struct SshClientA;
 
 #[async_trait]
-impl russh::client::Handler for SshClient {
+impl russh::client::Handler for SshClientA {
     type Error = anyhow::Error;
 
     async fn check_server_key(&mut self, _key: &ssh_key::PublicKey) -> Result<bool, Self::Error> {
@@ -110,7 +133,32 @@ impl russh::client::Handler for SshClient {
         _originator_port: u32,
         _session: &mut Session,
     ) -> Result<(), Self::Error> {
-        channel.data(&b"Hello, world!"[..]).await.unwrap();
+        channel.data(&b"A"[..]).await.unwrap();
+        channel.eof().await.unwrap();
+        Ok(())
+    }
+}
+
+struct SshClientB;
+
+#[async_trait]
+impl russh::client::Handler for SshClientB {
+    type Error = anyhow::Error;
+
+    async fn check_server_key(&mut self, _key: &ssh_key::PublicKey) -> Result<bool, Self::Error> {
+        Ok(true)
+    }
+
+    async fn server_channel_open_forwarded_tcpip(
+        &mut self,
+        channel: Channel<Msg>,
+        _connected_address: &str,
+        _connected_port: u32,
+        _originator_address: &str,
+        _originator_port: u32,
+        _session: &mut Session,
+    ) -> Result<(), Self::Error> {
+        channel.data(&b"B"[..]).await.unwrap();
         channel.eof().await.unwrap();
         Ok(())
     }