+++ tests/lib_configure_from_scratch.rs
@@ -53,6 +53,7 @@ async fn lib_configure_from_scratch() {
         load_balancing: LoadBalancing::Allow,
         allow_provided_subdomains: false,
         allow_requested_ports: true,
+        quota_per_user: None,
         random_subdomain_seed: None,
         txt_record_prefix: "_sandhole".into(),
         idle_connection_timeout: Duration::from_secs(1),