+++ tests/ssh_proxy_jump.rs
@@ -46,6 +46,7 @@ async fn ssh_proxy_jump() {
         load_balancing: LoadBalancing::Allow,
         allow_provided_subdomains: false,
         allow_requested_ports: false,
+        quota_per_user: None,
         random_subdomain_seed: None,
         txt_record_prefix: "_sandhole".into(),
         idle_connection_timeout: Duration::from_secs(2),