+++ tests/tcp_bind_random_ports.rs
@@ -40,6 +40,7 @@ async fn tcp_bind_random_ports() {
         load_balancing: LoadBalancing::Allow,
         allow_provided_subdomains: false,
         allow_requested_ports: false,
+        quota_per_user: None,
         random_subdomain_seed: None,
         txt_record_prefix: "_sandhole".into(),
         idle_connection_timeout: Duration::from_secs(1),