+++ tests/tcp_reject_low_ports.rs
@@ -39,6 +39,7 @@ async fn tcp_reject_low_ports() {
         load_balancing: LoadBalancing::Allow,
         allow_provided_subdomains: false,
         allow_requested_ports: true,
+        quota_per_user: None,
         random_subdomain_seed: None,
         txt_record_prefix: "_sandhole".into(),
         idle_connection_timeout: Duration::from_secs(1),